Add scan data panorama module
Adds a module to create scan data panorama images.
It uses the index bounds defined in the E57 file.
Includes a small executable to save to an JPG file.
The executable can be called with a path, the Data3D
GUID and an output path.

In the main application it is integrated in the context
menu of the Data3D.